scripts/ingestion/requirements.txt
recipe-scrapers==15.11.0
duckduckgo-search==8.1.1
google-api-python-client==2.194.0
supabase==2.28.3
requests==2.33.1
pytest==9.0.3
python-dotenv==1.2.2
